refactor: 순환 참조 문제로 인해 스토어 파일에서 type 분리
