+++ src/lib/stores/mandalaStore.ts
@@ -1,11 +1,5 @@
 import { create } from "zustand";
 import { immer } from "zustand/middleware/immer";
-import {
-  emptyDummyData,
-  toFlatStructure,
-  toggleStatus,
-  toLegacyStructure,
-} from "@/feature/mandala/service";
 import { findIdIndex, findKeyByValue } from "@/feature/mandala/utills/\bindex";
 import { persist } from "zustand/middleware";
 import { parseCellId } from "@/feature/mandala/service/parseCellId";
@@ -18,42 +12,20 @@ import type {
   MandalaFlatType,
   ServerMandalaType,
 } from "@/feature/mandala/service/type";
-
-export type Status = "DONE" | "UNDONE";
-export type EditingContext = "main" | "full" | "sub" | null;
-export type CancelReason = "blur" | "escape" | "enter";
-
-export type MandalaType<T = string> = {
-  core: {
-    goalId: T;
-    content: string;
-    originalId?: number | undefined;
-    status: Status;
-    mains: MainGoal<T>[];
-  };
-};
-export type MainGoal<T = string> = {
-  goalId: T;
-  originalId?: number | undefined; // 서버 원본 ID
-  position: number;
-  content: string;
-  status: Status;
-  subs: SubGoal<T>[];
-};
-
-export type SubGoal<T = string> = {
-  goalId: T;
-  originalId?: number | undefined; // 서버 원본 ID
-  position: number;
-  content: string;
-  status: Status;
-};
-
-export type DataOption = {
-  reminderEnabled: boolean;
-  remindInterval: string;
-  remindScheduledAt: string | null;
-};
+import {
+  emptyDummyData,
+  toFlatStructure,
+  toLegacyStructure,
+} from "@/feature/mandala/service/transform";
+import { toggleStatus } from "@/feature/mandala/service/status";
+import type {
+  DataOption,
+  EditingContext,
+  MainGoal,
+  MandalaType,
+  SubGoal,
+} from "./types/mandalart";
+import type { CancelReason } from "vitest";
 
 /**
  * @deprecated UI 렌더링은 flatData 사용
@@ -82,7 +54,7 @@ type States = {
 } & PersistedState;
 
 type PersistedState = {
-  data: MandalaType;
+  // data: MandalaType;
 };
 
 type Actions = {
@@ -437,6 +409,7 @@ export const useMandalaStore = create<States & Actions>()(
           } = persistedState;
           return rest;
         }
+
         return persistedState;
       },
     }
